Fixed addon,xml
I fixed missing reference in addon,xml and put the option to turn off
menu directly back into settings.xml. I am unsure that is needed with
the logic introduced with the "video_choice" menu option. With that,
there are four choices: 1) use a default profile, 2) switch based on
aspect ratio, 3) show a menu, or 4) turn off the leds.